+++ Ranger/UI/TourneyTable/TourneyTableView/Cells/Finishes/FinishesGraphCellView.swift
@@ -36,7 +36,7 @@ class FinishesGraphCellView: PlayerViewModelCellView
         setNeedsDisplay(self.bounds)
     }
     
-    override func setup(with playerViewModel: Player, in tableColumn: NSTableColumn?)
+    override func setup(with playerViewModel: Model.Player, in tableColumn: NSTableColumn?)
     {
         // Checks.
         guard let column = tableColumn else { return }